+++ CHANGELOG.md
@@ -6,13 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added:
+- Functionality to look up if a record has been restricted by DR by production ID.  
+
+### Changed:
+- Changed configuration naming for YAML key `holdback.dr` to `dr.holdback` in preparation for addition of `dr.restrictions`.
+
 ## [2.1.6](https://github.com/kb-dk/ds-present/releases/tag/ds-present-2.1.6) 2024-11-05
 ### Added
 - Added a more thorough parsing of categories into pretty genres.
 
 ### Changed 
 - solr 1.7.6: Changes related to search ranking.
-- Changed configuration naming for YAML key `holdback.dr` to `dr.holdback` in preparation for addition of `dr.restrictions`.
 
 ### Removed
 - Removed Schema.org field `conditionsOfAccess` and solr field `conditions_of_access` as this field was never used and the DR rights management couldn't be done in a single value