+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added:
-- Functionality to look up if a record has been restricted by DR by production ID.  
+- Functionality to look up if a record has been restricted by DR by production ID.
+- Solr 1.7.7: Add boolean field to show if record is restricted by DR.
 
 ### Changed:
 - Changed configuration naming for YAML key `holdback.dr` to `dr.holdback` in preparation for addition of `dr.restrictions`.